standardize json export to just a json array making export also compatible with table import from json array

+++ src/main/java/com/cannestro/drafttable/core/outbound/Structure.java
@@ -8,7 +8,7 @@ import java.util.Map;
 
 public record Structure(@NonNull String columnName,
                         @NonNull String type,
-                        double nullCount) implements Mappable {
+                        int nullCount) implements Mappable {
 
     @Override
     public Map<String, ?> asMap() {
